Fix CPAN testing failure
The directory locating test was failing on CPAN testers because it
was looking in TMPDIR before current dir. Ordinarily this is fine
but in some test environments TMPDIR is set to the working tree
so finds the directory in the wrong place. Reverse the entries
in the MYPATH environment variable.